Relatório Unificado - mensal e anual/requirements.txt
brotlipy==0.7.0
cachetools==5.3.1
chardet==5.2.0
contourpy==1.2.0
cycler==0.12.1
dnspython==2.4.2
et-xmlfile==1.1.0
EZSheets==2023.8.11
fonttools==4.44.0
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
httplib2==0.22.0
kiwisolver==1.4.5
matplotlib==3.8.1
mysql-connector==2.2.9
nbconvert
numpy==1.25.2
oauthlib==3.2.2
openpyxl==3.1.2
pandas==2.1.0
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pymongo==4.5.0
pyodbc==5.0.1
pyparsing==3.1.1
pywin32==305
referencing==0.30.2
requests-oauthlib==1.3.1
reportlab
rpds-py==0.10.0
rsa==4.9
soupsieve==2.4.1
tzdata==2023.3
uritemplate==4.1.1
webencodings==0.5.1